Several refactorings to testing
