+++ tests/core/pyspec/eth2spec/test/utils/kzg_tests.py
@@ -3,7 +3,7 @@ from eth_utils import (
     int_to_big_endian,
 )
 
-from eth2spec.fulu import spec
+from eth2spec.fulu import mainnet as spec
 from eth2spec.utils import bls
 
 ###############################################################################
